add MNIST code to INA org
