+++ environment.yml
@@ -1,0 +1,43 @@
+name: MNIST
+channels:
+  - conda-forge
+  - bioconda
+  - defaults
+dependencies:
+  - _libgcc_mutex=0.1=conda_forge
+  - _openmp_mutex=4.5=2_gnu
+  - bzip2=1.0.8=hda65f42_8
+  - ca-certificates=2025.8.3=hbd8a1cb_0
+  - ld_impl_linux-64=2.44=ha97dd6f_2
+  - libexpat=2.7.1=hecca717_0
+  - libffi=3.4.6=h2dba641_1
+  - libgcc=15.1.0=h767d61c_5
+  - libgomp=15.1.0=h767d61c_5
+  - liblzma=5.8.1=hb9d3cd8_2
+  - libmpdec=4.0.0=hb9d3cd8_0
+  - libsqlite=3.50.4=h0c1763c_0
+  - libuuid=2.41.2=he9a06e4_0
+  - libzlib=1.3.1=hb9d3cd8_2
+  - ncurses=6.5=h2d0b736_3
+  - openssl=3.5.3=h26f9b46_1
+  - pip=25.2=pyh145f28c_0
+  - python=3.13.7=h2b335a9_100_cp313
+  - python_abi=3.13=8_cp313
+  - readline=8.2=h8c095d6_2
+  - tk=8.6.13=noxft_hd72426e_102
+  - pip:
+      - contourpy==1.3.3
+      - cycler==0.12.1
+      - fonttools==4.60.0
+      - kiwisolver==1.4.9
+      - matplotlib==3.10.6
+      - numpy==2.3.3
+      - packaging==25.0
+      - pandas==2.3.2
+      - pillow==11.3.0
+      - pyparsing==3.2.5
+      - python-dateutil==2.9.0.post0
+      - pytz==2025.2
+      - seaborn==0.13.2
+      - six==1.17.0
+      - tzdata==2025.2
