+++ model/fashion-mnist/nn_parameters.npz
